Feito ?

+++ back-adonis/app/controllers/cursos_controller.ts
@@ -1,13 +1,21 @@
 import type { HttpContext } from '@adonisjs/core/http'
 import Curso from '#models/curso'
 import { createCurso, updateCurso } from '#validators/curso'
+import CursoPolicy from '#policies/curso_policy'
 
 export default class CursosController {
   /**
-   * Display a list of resources
+   * Display a list of resource
    */
-  async index({ response }: HttpContext) {
+  async index({ auth, bouncer, response }: HttpContext) {
     try {
+      // Usuário Autenticado
+      const user = auth.getUserOrFail()
+      // Verificar se o usuário pode listar posts
+      if (await bouncer.with(CursoPolicy).denies('list')) {
+        return response.forbidden({ message: 'Você não tem permissão para listar curso' })
+      }
+
       // const cursos = await Curso.all()
       const cursos = await Curso.query().preload('disciplinas').preload('alunos')
 
@@ -30,12 +38,19 @@ export default class CursosController {
   /**
    * Handle form submission for the create action
    */
-  async store({ request, response }: HttpContext) {
+  async store({ auth, bouncer, request, response }: HttpContext) {
     const payload = await request.validateUsing(createCurso)
-
     try {
-      const curso = await Curso.create({ ...payload })
+      // Usuário Autenticado
+      const user = auth.getUserOrFail()
+      // Verificar se o usuário pode listar posts
+      if (await bouncer.with(CursoPolicy).denies('create')) {
+        return response.forbidden({ message: 'Você não tem permissão para criar curso' })
+      }
 
+      const curso = await Curso.create({
+        ...payload,
+      })
       return response.status(201).json({
         message: 'OK',
         data: curso,
@@ -50,15 +65,21 @@ export default class CursosController {
   /**
    * Show individual record
    */
-  async show({ params, response }: HttpContext) {
+  async show({ auth, bouncer, params, response }: HttpContext) {
     try {
+      // Usuário Autenticado
+      const user = auth.getUserOrFail()
+      // Verificar se o usuário pode listar posts
+      if (await bouncer.with(CursoPolicy).denies('view')) {
+        return response.forbidden({ message: 'Você não tem permissão para ver curso' })
+      }
+
       // const curso = await Curso.findOrFail(params.id)
       const curso = await Curso.query()
         .where('id', params.id)
         .preload('disciplinas')
         .preload('alunos')
         .firstOrFail()
-
       return response.status(200).json({
         message: 'OK',
         data: curso,
@@ -73,15 +94,21 @@ export default class CursosController {
   /**
    * Edit individual record
    */
-  // async edit({ params }: HttpContext) {}
+  async edit({ params }: HttpContext) {}
 
   /**
    * Handle form submission for the edit action
    */
-  async update({ params, request, response }: HttpContext) {
+  async update({ auth, bouncer, params, request, response }: HttpContext) {
     const payload = await request.validateUsing(updateCurso)
-
     try {
+      // Usuário Autenticado
+      const user = auth.getUserOrFail()
+      // Verificar se o usuário pode listar posts
+      if (await bouncer.with(CursoPolicy).denies('edit')) {
+        return response.forbidden({ message: 'Você não tem permissão para alterar curso' })
+      }
+
       const curso = await Curso.findOrFail(params.id)
       await curso.merge({ ...payload }).save()
 
@@ -99,8 +126,15 @@ export default class CursosController {
   /**
    * Delete record
    */
-  async destroy({ params, response }: HttpContext) {
+  async destroy({ auth, bouncer, params, response }: HttpContext) {
     try {
+      // Usuário Autenticado
+      const user = auth.getUserOrFail()
+      // Verificar se o usuário pode listar posts
+      if (await bouncer.with(CursoPolicy).denies('delete')) {
+        return response.forbidden({ message: 'Você não tem permissão para remover curso' })
+      }
+
       const curso = await Curso.findOrFail(params.id)
       await curso.delete()
